requirements.txt
Pillow==3.1.0
Flask==0.10.1